payment-service/requirements.txt
Flask
psycopg2-binary
stripe